backend/requirements.txt
annotated-doc==0.0.4
annotated-types==0.7.0
anyio==4.12.0
av==16.0.1
certifi==2025.11.12
charset-normalizer==3.4.4
click==8.3.1
coloredlogs==15.0.1
ctranslate2==4.6.2
distro==1.9.0
fastapi==0.128.0
faster-whisper==1.2.1
filelock==3.20.1
flatbuffers==25.12.19
fsspec==2025.12.0
h11==0.16.0
hf-xet==1.2.0
httpcore==1.0.9
httpx==0.28.1
huggingface_hub==1.2.3
humanfriendly==10.0
idna==3.11
jiter==0.12.0
langdetect==1.0.9
mpmath==1.3.0
numpy==2.4.0
onnxruntime==1.23.2
openai==2.14.0
packaging==25.0
protobuf==6.33.2
pydantic==2.12.5
pydantic_core==2.41.5
python-dotenv==1.2.1
python-multipart==0.0.21
PyYAML==6.0.3
requests==2.32.5
setuptools==80.9.0
shellingham==1.5.4
six==1.17.0
sniffio==1.3.1
starlette==0.50.0
sympy==1.14.0
tokenizers==0.22.1
tqdm==4.67.1
typer-slim==0.21.0
typing-inspection==0.4.2
typing_extensions==4.15.0
urllib3==2.6.2
uvicorn==0.40.0
